backend/requirements.txt
uvicorn
fastapi
python-dotenv
colorama
ollama
chromadb
pytest
aiohttp
openai
groq
google-generativeai
# installed but not used
# pydantic

